[FEAT] #17 파일 이름 변경(discovery->discoveryFilm, home->discovery, mainPage->home)
