+++ src/App.jsx
@@ -2,19 +2,19 @@ import React from 'react';
 import MovieDetailPage from './pages/MovieDetailPage.jsx';
 import { BrowserRouter, Route, Routes } from "react-router-dom"
 import Login from "./pages/login"
-import Home from "./pages/Home"
+import Discovery from "./pages/Discovery.jsx"
 import RankingPage from "./pages/RankingPage"
 import MyPage from './pages/MyPage.jsx';
 import MyPageHistroy from './pages/MyPageHistroy.jsx';
 import MovieDiscoveryPage from "./pages/MovieDiscoveryPage"
-import MainPage from './pages/MainPage.jsx';
+import Home from './pages/Home.jsx';
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/home" element={<MainPage />} />
-        <Route path="/discovery" element={<Home />} />
+        <Route path="/home" element={<Home />} />
+        <Route path="/discovery" element={<Discovery />} />
         <Route path="/discovery/discoveryfilm" element={<MovieDiscoveryPage />} />
         <Route path="/login" element={<Login />} />
         <Route path="/movie" element={<MovieDetailPage />} />
