+++ src/components/MovieList.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from 'react-router-dom';
-import "../styles/home/movieList.css";
+import "../styles/discovery/movieList.css";
 
 export default function MovieList(props) {
     return (
