+++ src/styles/discovery/home.css
@@ -9,6 +9,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    padding-bottom: 50px;
 }
 
 .home-recResult{
