+++ src/styles/discovery/movieList.css
