+++ src/styles/discovery/recReview.css
@@ -70,13 +70,14 @@
                     font-family: "Freesentation-7Bold";
                     font-size: 11.7px;
                     color: #FFFFFF;
-                    margin-right: 3px;
+                    margin-right: 5px;
                 }
                 .profile-img{
                     width: 19.5px;
                     height: 19.5px;
                     border-radius: 50%;
                     background-size: cover;
+                    background-color: #FFFFFF;
                 }
             }
         }
