+++ src/styles/discoveryFilm/DiscoveryMovie.css
