+++ src/styles/discoveryFilm/DiscoveryResult.css
